Unit test updates

+++ sample/src/main/java/com/afollestad/recyclicalsample/MainActivity.kt
@@ -71,7 +71,7 @@ class MainActivity : AppCompatActivity() {
       withDataSource(dataSource)
 
       withItem<MyListItem>(R.layout.my_list_item) {
-        hasStableIds { it.id.toLong() }
+        hasStableIds { it.id }
 
         onBind(::MyViewHolder) { _, item ->
           icon.setImageResource(
